+++ CHANGES.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for function calls.
 - Support for the `buff`, `principal`, `response`, and `tuple` types.
 - Support for `is-ok`, `is-err`, and `err`.
+- Support for `try!` and all `unwrap*` forms.
 
 ## [0.2.0] - 2020-09-22